+++ src/my_first_adventure_game/game/entities/__init__.py
@@ -1,3 +1,8 @@
+from my_first_adventure_game.game.entities.caretaker_behavior import (
+    CARETAKER_SIDESTEP_TARGET_ID,
+    CaretakerBehavior,
+    CaretakerPhase,
+)
 from my_first_adventure_game.game.entities.caretaker_movement import (
     caretaker_sidestep_target,
 )
@@ -8,6 +13,9 @@ from my_first_adventure_game.game.entities.player import Player
 from my_first_adventure_game.game.entities.wall_stain import WallStain
 
 __all__ = [
+    "CARETAKER_SIDESTEP_TARGET_ID",
+    "CaretakerBehavior",
+    "CaretakerPhase",
     "Enemy",
     "NPC",
     "Player",